Big refactoring of TRequest
The class is now taking advantages of the unified error messages for
invalid/expired token.
The FormRequest and request methods have been merged in a single method request
capable of handling json and multi-part/form-data bodies.
The method will automatically convert the response body in one of the following according
to the Content-Type header :
- text/json
- text/plain
- text/HTML
- Blob

Refer to the documentation for mor informations and examples.