TRequest/Application
The refreshToken() method is now a part of the Application class as it makes more
sense.

The method checkAccessTokenValidity() has been added to the Application class.

The TRequest.request() method now checks locally if the access token has not expired (
and try to refresh it otherwise) BEFORE  trying to execute the request.